refactor: context comment/cleanup
